+++ generators/app/templates/javascript/src/components/missing/Missing.jsx
@@ -7,7 +7,10 @@ const Missing = () => (
   <Layout>
     <SectionHeader title="That's a 404" description="You must be lost..." />
     <p>
-      <Link to="/">Click here</Link> to return to the main application.
+      <Link to="/" className="link-purple">
+        Click here
+      </Link>{' '}
+      to return to the main application.
     </p>
   </Layout>
 );